Move tests/base_test content to pipobot/lib/module_test so it can be imported 'outside' pipobot (to unit-test modules)
