+++ pipobot/lib/module_test.py
@@ -1,9 +1,17 @@
 # -*- coding: utf-8 -*-
+import queue
 import random
 import re
 import string
 import unittest
+from sqlalchemy import create_engine
+from sqlalchemy.orm import sessionmaker
 
+from pipobot.bot_test import TestBot
+from pipobot.lib.bdd import Base
+
+import builtins
+builtins._ = lambda x: x
 
 class ModuleTest(unittest.TestCase):
     __usable = False
@@ -11,15 +19,6 @@ class ModuleTest(unittest.TestCase):
     def __init__(self, methodName='runTest', bot=None):
         unittest.TestCase.__init__(self, methodName)
         self.bot = bot
-
-    @staticmethod
-    def parametrize(testcase_class, bot):
-        testloader = unittest.TestLoader()
-        testnames = testloader.getTestCaseNames(testcase_class)
-        suite = unittest.TestSuite()
-        for name in testnames:
-            suite.addTest(testcase_class(name, bot))
-        return suite
 
     def bot_answer(self, input_msg, user="test"):
         th = self.bot.create_msg(user, input_msg)
@@ -40,3 +39,41 @@ class ModuleTest(unittest.TestCase):
 
 def string_gen(size):
     return "".join([random.choice(string.ascii_lowercase) for i in range(size)])
+
+
+DOMAIN = "domain.tld"
+CHAN = "test@%s" % DOMAIN
+
+
+class FakeUser:
+    def __init__(self, name, bot, tester, jid=None, role="moderator", chan=CHAN):
+        self.name = name
+        self.bot = bot
+        if jid is None:
+            jid = "%s@%s" % (name, DOMAIN)
+        self.jid = jid
+        self.role = role
+        self.chan = chan
+        self.tester = tester
+        self.bot.occupants.add_user(name, jid, role)
+
+    def ask(self, msg, expected):
+        ret = self.bot.create_msg(self.name, msg)
+        ret.join()
+        ret = self.bot.output.get()
+        self.tester.assertEqual(ret, expected)
+
+    def change_nick(self, new_nick):
+        self.bot.users.rm_user(self.name)
+        self.name = new_nick
+        self.bot.occupants.add_user(new_nick, self.jid, self.role)
+
+    def leave(self):
+        self.bot.users.rm_user(self.name)
+
+def create_test_bot(mods):
+    engine = create_engine('sqlite:///:memory:')
+    Session = sessionmaker(bind=engine)
+    session = Session()
+    Base.metadata.create_all(engine)
+    return TestBot("pipotest", "login", CHAN, mods, session, queue.Queue())
